requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1 